requirements.txt
streamlit
llama_index.llms
llama_index
docx2txt
llama-cpp-python
llama_index
io
os
boto3